requirements.txt
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1737352602016/work
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
black==25.1.0
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1733827268945/work
boto3==1.28.57
botocore==1.31.85
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1725267563793/work
cachetools==5.5.1
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560578465/work
cfgv==3.4.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1735929714516/work
chromedriver-py==135.0.7049.42
click==8.1.8
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
conda @ file:///Users/runner/miniforge3/conda-bld/conda_1736460303265/work
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1727359833193/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1736345463896/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1741620732069/work
cryptography==44.0.2
distlib==0.3.9
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1734729835256/work
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.5.0
django-bootstrap5==25.1
django-cors-headers==4.7.0
django-filter==25.1
django-storages==1.13.2
djangorestframework==3.15.2
filelock==3.18.0
frozendict @ file:///Users/runner/miniforge3/conda-bld/frozendict_1728841362247/work
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.67.0
gunicorn==23.0.0
h11==0.14.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httplib2==0.22.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
identify==2.6.9
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
jmespath==1.0.1
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1733814567314/work
jsonpointer @ file:///Users/runner/miniforge3/conda-bld/jsonpointer_1725302931775/work
jwcrypto==1.5.6
libmambapy @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1735835787728/work/libmambapy
mamba @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1735835787728/work/mamba
Markdown==3.7
menuinst @ file:///Users/runner/miniforge3/conda-bld/menuinst_1731147022576/work
mypy-extensions==1.0.0
nodeenv==1.9.1
oauthlib==3.2.2
outcome==1.3.0.post0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pathspec==0.12.1
pillow==11.1.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1733222765875/work
pre_commit==4.1.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg==3.2.4
psycopg-binary==3.2.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1732588505673/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
PyJWT==2.10.1
pyparsing==3.2.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
rsa==4.9
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1736248058782/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1728724504988/work
s3transfer==0.7.0
selenium==4.30.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
trio==0.29.0
trio-websocket==0.12.2
truststore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_truststore_1739009763/work
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.29.3
websocket-client==1.8.0
wheel==0.45.1
whitenoise==6.9.0
wsproto==1.2.0
zstandard==0.23.0
